test fancinette reussi

+++ libftprintf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: msuter <msuter@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/15 14:42:59 by msuter            #+#    #+#             */
-/*   Updated: 2025/10/20 09:54:54 by msuter           ###   ########.fr       */
+/*   Updated: 2025/10/20 13:21:54 by msuter           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_printf(const char *format, ...);
 int		ft_putchar(char c);
 int		ft_putstr(char *str);
 int		ft_putnbr_base(unsigned long nb);
-int		modif_putnbr(unsigned int n);
+int		modif_putnbr(unsigned long n);
 int		ft_putnbr_base_maj(unsigned long nb);
 
 #endif
